requirements.txt
fastapi[all]
uvicorn
python-docx
pdfkit
pydantic
PyPDF2
python-multipart
reportlab
docx2pdf
pdf2docx